+++ libfreerdp/core/capabilities.c
@@ -2977,7 +2977,7 @@ BOOL rdp_print_capability_sets(wStream* s, UINT16 numberCapabilities, BOOL recei
 
 		em = bm + length;
 
-		if (Stream_GetRemainingLength(s) < length - 4)
+		if (Stream_GetRemainingLength(s) < (size_t) (length - 4))
 		{
 			fprintf(stderr, "error processing stream\n");
 			return FALSE;
@@ -3176,7 +3176,7 @@ BOOL rdp_read_capability_sets(wStream* s, rdpSettings* settings, UINT16 numberCa
 
 		em = bm + length;
 
-		if (Stream_GetRemainingLength(s) < length - 4)
+		if (Stream_GetRemainingLength(s) < ((size_t) length - 4))
 		{
 			fprintf(stderr, "error processing stream\n");
 			return FALSE;
@@ -3555,7 +3555,7 @@ BOOL rdp_recv_confirm_active(rdpRdp* rdp, wStream* s)
 	Stream_Read_UINT16(s, lengthSourceDescriptor); /* lengthSourceDescriptor (2 bytes) */
 	Stream_Read_UINT16(s, lengthCombinedCapabilities); /* lengthCombinedCapabilities (2 bytes) */
 
-	if (Stream_GetRemainingLength(s) < lengthSourceDescriptor + 4)
+	if (((int) Stream_GetRemainingLength(s)) < lengthSourceDescriptor + 4)
 		return FALSE;
 
 	Stream_Seek(s, lengthSourceDescriptor); /* sourceDescriptor */
